Move find-or-create-publication for ingest to mixin, so it can be used in other ingests, not only NDNP.
